api/requirements.txt
absl-py==1.0.0
astunparse==1.6.3
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
BareNecessities==0.2.8
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
cachetools==4.2.4
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
click==8.0.1
colorama==0.4.4
coverage==6.3.2
elasticsearch==7.13.4
finnhub-python==2.3.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Mail==0.9.1
flatbuffers==2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.6.0
idna==2.10
importlib-metadata==4.8.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
lxml==4.6.2
Mail==2.1.0
Markdown==3.3.6
MarkupSafe==2.0.1
multitasking==0.0.9
mysql-connector-python==8.0.25
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.2.1
pluggy==1.0.0
protobuf==3.17.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.0.1
pytest-cov==3.0.0
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.15.0
soupsieve==2.2.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
tk==0.1.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.0.1
urllib3==1.26.3
Werkzeug==2.0.1
wikipedia==1.4.0
wrapt==1.13.3
yfinance==0.1.55
zipp==3.6.0
